added authorization, delete socket

+++ server/controllers/messageController.ts
@@ -1,14 +1,15 @@
 import { Router } from 'express';
 import Message from '../models/message.model';
+import { authenticateToken } from './authenticationController';
 
 const router = Router();
 
 // CREATING NEW MESSAGE
-router.post('/', async (req, res) => {
+router.post('/', authenticateToken, async (req, res) => {
   try {
     const message = new Message({
       roomId: req.body.roomId,
-      from: req.body.from,
+      from: req.body.authId,
       message: req.body.message,
     });
 
@@ -27,14 +28,20 @@ router.post('/', async (req, res) => {
   }
 });
 
-router.delete('/:roomId', async (req, res) => {
+router.delete('/:roomId', authenticateToken, async (req, res) => {
   try {
     const { roomId } = req.params;
+    const { authId } = req.body;
 
-    const messages = await Message.deleteMany({ roomId });
-    res.json({
-      msg: `Deleted room: ${roomId}. Total messages: ${messages.deletedCount}`,
-    });
+    const ids = roomId.split('_');
+
+    if (ids.includes(authId)) {
+      const messages = await Message.deleteMany({ roomId });
+      return res.json({
+        msg: `Deleted room: ${roomId}. Total messages: ${messages.deletedCount}`,
+      });
+    }
+    res.sendStatus(403);
   } catch (err) {
     const msg = (err as Error).message;
     if (msg) return res.status(400).send({ error: msg });
@@ -43,10 +50,17 @@ router.delete('/:roomId', async (req, res) => {
 });
 
 // GETTING ALL MESSEGES BY ROOM ID
-router.get('/:roomId', async (req, res) => {
+router.get('/:roomId', authenticateToken, async (req, res) => {
   try {
-    const messages = await Message.find({ roomId: req.params.roomId });
-    res.json(messages);
+    const { roomId } = req.params;
+    const { authId } = req.body;
+
+    const ids = roomId.split('_');
+    if (ids.includes(authId)) {
+      const messages = await Message.find({ roomId: roomId });
+      return res.json(messages);
+    }
+    res.sendStatus(403);
   } catch (err) {
     const msg = (err as Error).message;
     if (msg) return res.status(400).send({ error: msg });
